requirements.txt
pip
asyncio
discord.py
requests
boticordpy
yarl
